+++ ReleaseNotes.md
@@ -7,6 +7,7 @@
 * SBBRadioButton and SBBCheckbox now reflect theming (color) changes.
 * Breaking change: Renaming of init parameters of SBBRadioButton and SBBCheckbox according to newest Figma version. 
 * Breaking change: SBBRadioButton and SBBCheckbox now feature an optional subtitle instead of customizable content.
+* Breaking change: Renaming of SBBInfoViewCollapsible to SBBAccordion.
 
 ## V0.5.2 - 12 November 2021